clean up C4244 warnings on MSVC builds
C4244 appertains to implicit truncation of integer values [1].
Theoretically, these warnings could be detected using
`-fsanitize=implicit-unsigned-integer-truncation` and
`-fsanitize=implicit-signed-integer-truncation` as well as statically
using `-Werror=implicit-int-conversion`.

[1] https://learn.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-levels-3-and-4-c4244?view=msvc-170

+++ src/node.c
@@ -348,7 +348,7 @@ int cmark_node_set_heading_level(cmark_node *node, int level) {
 
   switch (node->type) {
   case CMARK_NODE_HEADING:
-    node->as.heading.level = level;
+    node->as.heading.level = (int8_t)level;
     return 1;
 
   default:
